Fix cleanup for K8s dependencies
