+++ ocw/lib/emailnotify.py
@@ -4,7 +4,7 @@ import logging
 from email.mime.text import MIMEText
 from texttable import Texttable
 from django.urls import reverse
-from webui.settings import PCWConfig
+from webui.PCWConfig import PCWConfig
 from webui.settings import build_absolute_uri
 from ..models import Instance
