+++ tests/conftest.py
@@ -12,7 +12,7 @@ def set_pcw_ini(filename, add=''):
 @pytest.fixture(autouse=True)
 def pcw_file():
     tmpFile = tempfile.mkstemp()
-    webui.settings.CONFIG_FILE = tmpFile[1]
+    webui.PCWConfig.CONFIG_FILE = tmpFile[1]
     set_pcw_ini(tmpFile[1])
     yield tmpFile[1]
     if os.path.exists(tmpFile[1]):
