+++ tests/test_ec2.py
@@ -1,7 +1,7 @@
 import os
 import kubernetes
 from ocw.lib.EC2 import EC2, Provider
-from webui.settings import PCWConfig
+from webui.PCWConfig import PCWConfig
 from tests.generators import mock_get_feature_property
 from tests.generators import ec2_max_age_days
 from faker import Faker
